backends/account-back/requirements.txt
Django==5.2.5
django-ninja==1.4.3
django-ninja-jwt[crypto]==5.3.7
django-redis==6.0.0
redis==6.2.0
psycopg==3.2.9
uvicorn==0.35.0
gunicorn==23.0.0
whitenoise==6.9.0